# simple-crud-api/api-spec/openapi.yaml
openapi: 3.0.3
info:
  title: Simple CRUD API
  description: A simple CRUD API for managing items - Contract First approach
  version: 1.0.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8080
    description: Local development server
paths:
  /api/items:
    get:
      summary: Get all items
      description: Retrieves a list of all items
      operationId: getAllItems
      tags:
        - items
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Item'
    post:
      summary: Create a new item
      description: Creates a new item
      operationId: createItem
      tags:
        - items
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateItemRequest'
      responses:
        '201':
          description: Item created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '400':
          description: Invalid input
  /api/items/{id}:
    get:
      summary: Get item by ID
      description: Retrieves a specific item by its ID
      operationId: getItemById
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
    put:
      summary: Update an item
      description: Updates an existing item
      operationId: updateItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateItemRequest'
      responses:
        '200':
          description: Item updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Item'
        '404':
          description: Item not found
    delete:
      summary: Delete an item
      description: Deletes an item by its ID
      operationId: deleteItem
      tags:
        - items
      parameters:
        - name: id
          in: path
          required: true
          description: The item ID
          schema:
            type: integer
            format: int64
      responses:
        '204':
          description: Item deleted successfully
        '404':
          description: Item not found
components:
  schemas:
    Item:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: Unique identifier for the item
          example: 1
        name:
          type: string
          description: Name of the item
          example: "Sample Item"
        description:
          type: string
          description: Description of the item
          example: "This is a sample item description"
        price:
          type: number
          format: double
          description: Price of the item
          example: 19.99
        createdAt:
          type: string
          format: date-time
          description: Creation timestamp
      required:
        - id
        - name
    CreateItemRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the item
          example: "New Item"
        description:
          type: string
          description: Description of the item
          example: "A new item description"
        price:
          type: number
          format: double
          description: Price of the item
          example: 29.99
      required:
        - name
    UpdateItemRequest:
      type: object
      properties:
        name:
          type: string
          description: Name of the item
          example: "Updated Item"
        description:
          type: string
          description: Description of the item
          example: "An updated item description"
        price:
          type: number
          format: double
          description: Price of the item
          example: 39.99